Modified addNumber() and added showNumbers()
This deploy modifies addNumber() passing the complete list by reference to the function.
Also creates showNumbers() with the complete list by reference and shows the numbers in the console.

Maybe in the future is better to avoid passing the List by reference and just initialize the List as a Global List in Program.cs. But at this time it's only laboratory so it's useful and simple to use it this way